+++ tests/test_suite.py
@@ -378,6 +378,42 @@ def test_opensubtitles_tvshow_missing_imdb_id():
 
     assert len(search.results) > 0
 
+def test_opensubtitles_tvshow_missing_imdb_id_but_in_url():
+    a4ksubtitles_api = api.A4kSubtitlesApi({'kodi': True})
+    __remove_all_cache(a4ksubtitles_api)
+
+    # search
+    settings = {
+        'opensubtitles.enabled': 'true',
+    }
+    video_meta = {
+        'imdb_id': '',
+        'season': '',
+        'episode': '',
+        'url': 'https://example.com/example.mkv?imdb_id=tt11610548'
+    }
+    search = __search_movie(a4ksubtitles_api, settings, video_meta)
+
+    assert len(search.results) > 0
+
+def test_opensubtitles_tvshow_missing_imdb_id_but_in_url_with_show_id_and_meta_for_ep_and_season():
+    a4ksubtitles_api = api.A4kSubtitlesApi({'kodi': True})
+    __remove_all_cache(a4ksubtitles_api)
+
+    # search
+    settings = {
+        'opensubtitles.enabled': 'true',
+    }
+    video_meta = {
+        'imdb_id': '',
+        'season': '',
+        'episode': '',
+        'url': 'https://example.com/example.mkv?imdb_id=tt10155688&season=1&episode=2'
+    }
+    search = __search_movie(a4ksubtitles_api, settings, video_meta)
+
+    assert len(search.results) > 0
+
 def test_bsplayer():
     a4ksubtitles_api = api.A4kSubtitlesApi({'kodi': True})
     __remove_all_cache(a4ksubtitles_api)